refactor(message): move filters from message.ts
Fixes #4487
