+++ src/bot/filters/online.ts
@@ -1,0 +1,14 @@
+import api from '../api';
+
+import { ResponseFilter } from '.';
+
+const online: ResponseFilter = {
+  '(onlineonly)': async function () {
+    return api.isStreamOnline;
+  },
+  '(offlineonly)': async function () {
+    return !(api.isStreamOnline);
+  },
+};
+
+export { online };
