+++ src/bot/filters/price.ts
@@ -1,0 +1,15 @@
+import { getRepository } from 'typeorm';
+
+import { Price } from '../database/entity/price';
+import points from '../systems/points';
+
+import type { ResponseFilter } from '.';
+
+const price: ResponseFilter = {
+  '(price)': async function (_variable, attr) {
+    const cmd = await getRepository(Price).findOne({ command: attr.cmd, enabled: true });
+    return [price, await points.getPointsName(cmd?.price ?? 0)].join(' ');
+  },
+};
+
+export { price };
